# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "FastAPI",
  "version": "0.1.0"
 },
 "paths": {
  "/v0/job/query": {
   "get": {
    "summary": "Get Jobs",
    "operationId": "get_jobs_v0_job_query_get",
    "parameters": [
     {
      "name": "cluster",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cluster"
      }
     },
     {
      "name": "job_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "array",
         "items": {
          "type": "string"
         }
        },
        {
         "type": "null"
        }
       ],
       "title": "Job Id"
      }
     },
     {
      "name": "job_state",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/SlurmState"
        },
        {
         "type": "null"
        }
       ],
       "title": "Job State"
      }
     },
     {
      "name": "username",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Username"
      }
     },
     {
      "name": "start",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date-time"
        },
        {
         "type": "null"
        }
       ],
       "title": "Start"
      }
     },
     {
      "name": "end",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date-time"
        },
        {
         "type": "null"
        }
       ],
       "title": "End"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "string"
         },
         "title": "Response Get Jobs V0 Job Query Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v0/job/list": {
   "get": {
    "summary": "List Jobs",
    "operationId": "list_jobs_v0_job_list_get",
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "per_page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 100,
       "title": "Per Page"
      }
     },
     {
      "name": "cluster",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cluster"
      }
     },
     {
      "name": "job_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "array",
         "items": {
          "type": "string"
         }
        },
        {
         "type": "null"
        }
       ],
       "title": "Job Id"
      }
     },
     {
      "name": "job_state",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/SlurmState"
        },
        {
         "type": "null"
        }
       ],
       "title": "Job State"
      }
     },
     {
      "name": "username",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Username"
      }
     },
     {
      "name": "start",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date-time"
        },
        {
         "type": "null"
        }
       ],
       "title": "Start"
      }
     },
     {
      "name": "end",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date-time"
        },
        {
         "type": "null"
        }
       ],
       "title": "End"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SlurmJobList"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v0/job/count": {
   "get": {
    "summary": "Count Jobs",
    "operationId": "count_jobs_v0_job_count_get",
    "parameters": [
     {
      "name": "cluster",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cluster"
      }
     },
     {
      "name": "job_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "array",
         "items": {
          "type": "string"
         }
        },
        {
         "type": "null"
        }
       ],
       "title": "Job Id"
      }
     },
     {
      "name": "job_state",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/SlurmState"
        },
        {
         "type": "null"
        }
       ],
       "title": "Job State"
      }
     },
     {
      "name": "username",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Username"
      }
     },
     {
      "name": "start",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date-time"
        },
        {
         "type": "null"
        }
       ],
       "title": "Start"
      }
     },
     {
      "name": "end",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date-time"
        },
        {
         "type": "null"
        }
       ],
       "title": "End"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "integer",
         "title": "Response Count Jobs V0 Job Count Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v0/job/id/{oid}": {
   "get": {
    "summary": "Get Job",
    "operationId": "get_job_v0_job_id__oid__get",
    "parameters": [
     {
      "name": "oid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Oid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SlurmJob"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/v0/user/query": {
   "get": {
    "summary": "Query Users",
    "description": "Search users. Return user UUIDs.",
    "operationId": "query_users_v0_user_query_get",
    "parameters": [
     {
      "name": "display_name",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Display Name"
      }
     },
     {
      "name": "email",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Email"
      }
     },
     {
      "name": "member_type",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/MemberType"
        },
        {
         "type": "null"
        }
       ],
       "title": "Member Type"
      }
     },
     {
      "name": "member_start",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date-time"
        },
        {
         "type": "null"
        }
       ],
       "title": "Member Start"
      }
     },
     {
      "name": "member_end",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date-time"
        },
        {
         "type": "null"
        }
       ],
       "title": "Member End"
      }
     },
     {
      "name": "supervisor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid4"
        },
        {
         "type": "null"
        }
       ],
       "title": "Supervisor"
      }
     },
     {
      "name": "supervisor_start",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date-time"
        },
        {
         "type": "null"
        }
       ],
       "title": "Supervisor Start"
      }
     },
     {
      "name": "supervisor_end",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date-time"
        },
        {
         "type": "null"
        }
       ],
       "title": "Supervisor End"
      }
     },
     {
      "name": "co_supervisor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid4"
        },
        {
         "type": "null"
        }
       ],
       "title": "Co Supervisor"
      }
     },
     {
      "name": "co_supervisor_start",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date-time"
        },
        {
         "type": "null"
        }
       ],
       "title": "Co Supervisor Start"
      }
     },
     {
      "name": "co_supervisor_end",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date-time"
        },
        {
         "type": "null"
        }
       ],
       "title": "Co Supervisor End"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "string",
          "format": "uuid4"
         },
         "title": "Response Query Users V0 User Query Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v0/user/list": {
   "get": {
    "summary": "List Users",
    "operationId": "list_users_v0_user_list_get",
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "per_page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 100,
       "title": "Per Page"
      }
     },
     {
      "name": "display_name",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Display Name"
      }
     },
     {
      "name": "email",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Email"
      }
     },
     {
      "name": "member_type",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/MemberType"
        },
        {
         "type": "null"
        }
       ],
       "title": "Member Type"
      }
     },
     {
      "name": "member_start",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date-time"
        },
        {
         "type": "null"
        }
       ],
       "title": "Member Start"
      }
     },
     {
      "name": "member_end",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date-time"
        },
        {
         "type": "null"
        }
       ],
       "title": "Member End"
      }
     },
     {
      "name": "supervisor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid4"
        },
        {
         "type": "null"
        }
       ],
       "title": "Supervisor"
      }
     },
     {
      "name": "supervisor_start",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date-time"
        },
        {
         "type": "null"
        }
       ],
       "title": "Supervisor Start"
      }
     },
     {
      "name": "supervisor_end",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date-time"
        },
        {
         "type": "null"
        }
       ],
       "title": "Supervisor End"
      }
     },
     {
      "name": "co_supervisor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid4"
        },
        {
         "type": "null"
        }
       ],
       "title": "Co Supervisor"
      }
     },
     {
      "name": "co_supervisor_start",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date-time"
        },
        {
         "type": "null"
        }
       ],
       "title": "Co Supervisor Start"
      }
     },
     {
      "name": "co_supervisor_end",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date-time"
        },
        {
         "type": "null"
        }
       ],
       "title": "Co Supervisor End"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserList"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v0/user/id/{uuid}": {
   "get": {
    "summary": "Get User By Id",
    "description": "Get user with given UUID.",
    "operationId": "get_user_by_id_v0_user_id__uuid__get",
    "parameters": [
     {
      "name": "uuid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid4",
       "title": "Uuid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserData"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v0/user/email/{email}": {
   "get": {
    "summary": "Get User By Email",
    "description": "Get user with given email.",
    "operationId": "get_user_by_email_v0_user_email__email__get",
    "parameters": [
     {
      "name": "email",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Email"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserData"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "MemberType": {
    "type": "string",
    "enum": [
     "master",
     "phd",
     "postdoc",
     "professor",
     "staff",
     "intern"
    ],
    "title": "MemberType"
   },
   "SlurmJob": {
    "properties": {
     "id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Id"
     },
     "cluster_name": {
      "type": "string",
      "title": "Cluster Name"
     },
     "account": {
      "type": "string",
      "title": "Account"
     },
     "job_id": {
      "type": "integer",
      "title": "Job Id"
     },
     "array_job_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Array Job Id"
     },
     "task_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Task Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "user": {
      "type": "string",
      "title": "User"
     },
     "group": {
      "type": "string",
      "title": "Group"
     },
     "job_state": {
      "$ref": "#/components/schemas/SlurmState"
     },
     "exit_code": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Exit Code"
     },
     "signal": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Signal"
     },
     "partition": {
      "type": "string",
      "title": "Partition"
     },
     "nodes": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Nodes"
     },
     "work_dir": {
      "type": "string",
      "title": "Work Dir"
     },
     "constraints": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Constraints"
     },
     "priority": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Priority"
     },
     "qos": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Qos"
     },
     "CLEAR_SCHEDULING": {
      "type": "boolean",
      "title": "Clear Scheduling",
      "default": false
     },
     "STARTED_ON_SUBMIT": {
      "type": "boolean",
      "title": "Started On Submit",
      "default": false
     },
     "STARTED_ON_SCHEDULE": {
      "type": "boolean",
      "title": "Started On Schedule",
      "default": false
     },
     "STARTED_ON_BACKFILL": {
      "type": "boolean",
      "title": "Started On Backfill",
      "default": false
     },
     "time_limit": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Time Limit"
     },
     "submit_time": {
      "type": "string",
      "format": "date-time",
      "title": "Submit Time"
     },
     "start_time": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Start Time"
     },
     "end_time": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "End Time"
     },
     "elapsed_time": {
      "type": "number",
      "title": "Elapsed Time"
     },
     "latest_scraped_start": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Latest Scraped Start"
     },
     "latest_scraped_end": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Latest Scraped End"
     },
     "requested": {
      "$ref": "#/components/schemas/SlurmResources"
     },
     "allocated": {
      "$ref": "#/components/schemas/SlurmResources"
     },
     "stored_statistics": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/JobStatistics"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "cluster_name",
     "account",
     "job_id",
     "name",
     "user",
     "group",
     "job_state",
     "partition",
     "nodes",
     "work_dir",
     "submit_time",
     "elapsed_time",
     "requested",
     "allocated"
    ],
    "title": "SlurmJob",
    "description": "Holds data for a Slurm job."
   },
   "SlurmJobList": {
    "properties": {
     "jobs": {
      "items": {
       "$ref": "#/components/schemas/SlurmJob"
      },
      "type": "array",
      "title": "Jobs"
     },
     "page": {
      "type": "integer",
      "title": "Page"
     },
     "per_page": {
      "type": "integer",
      "title": "Per Page"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     }
    },
    "type": "object",
    "required": [
     "jobs",
     "page",
     "per_page",
     "total"
    ],
    "title": "SlurmJobList"
   },
...
   "SlurmState": {
    "type": "string",
    "enum": [
     "BOOT_FAIL",
     "CANCELLED",
     "COMPLETED",
     "CONFIGURING",
     "COMPLETING",
     "DEADLINE",
     "FAILED",
     "NODE_FAIL",
     "OUT_OF_MEMORY",
     "PENDING",
     "PREEMPTED",
     "RUNNING",
     "RESV_DEL_HOLD",
     "REQUEUE_FED",
     "REQUEUE_HOLD",
     "REQUEUED",
     "RESIZING",
     "REVOKED",
     "SIGNALING",
     "SPECIAL_EXIT",
     "STAGE_OUT",
     "STOPPED",
     "SUSPENDED",
     "TIMEOUT"
    ],
    "title": "SlurmState",
    "description": "Possible Slurm job states.\n\nReference: https://slurm.schedmd.com/squeue.html#SECTION_JOB-STATE-CODES"
   },
...
   "UserList": {
    "properties": {
     "users": {
      "items": {
       "$ref": "#/components/schemas/UserData"
      },
      "type": "array",
      "title": "Users"
     },
     "page": {
      "type": "integer",
      "title": "Page"
     },
     "per_page": {
      "type": "integer",
      "title": "Per Page"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     }
    },
    "type": "object",
    "required": [
     "users",
     "page",
     "per_page",
     "total"
    ],
    "title": "UserList"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}